requirements.txt
Flask==3.1.0
Werkzeug>=3.1.0,<4.0.0
gunicorn==23.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Flask-Compress==1.17
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
pillow==11.2.1
python-dotenv==1.1.0
python-slugify==8.0.4
python-socketio==5.12.1
psycopg2-binary==2.9.10
